Move threading::shutdown into threading::lifecycle
This move provides a location for future threading
lifecycle traits to be added.
